Revamped Term object. Completed 2. and 4. from the 'Things To Add' list
New regex patterns, and different method of breaking down the function.
About ready for basic calculator functionality. Each term is created as
a Term object with the following properties: (term, base, sign, coeff,
exp, variable). Term is mainly for display purposes, base is used to
derive sign and coeff, sign is either + or -, coeff is the numerical
coefficient and will always be a positive integer (need to add float
support), exp is the exponent (need to add support for floats), and
lastly variable is set at beginning of function evaluation and only
exists for future purposes.